+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add a default sorting option.
 
+### Changed
+
+- Change error error on status bar to be more prominent
+
 ## [0.9.3] - 2021-09-17
 
 ### Added